Getting things running with VS2019

+++ mediaLib/src/Media/mColourLookup.cpp
@@ -158,9 +158,9 @@ mFUNCTION(mColourLookup_At, mPtr<mColourLookup> &colourLookup, const mVec3f posi
 
   mVec3f pos = mVec3f(mClamp(position.x, 0.f, 1.f), mClamp(position.y, 0.f, 1.f), mClamp(position.z, 0.f, 1.f)) * (mVec3f(colourLookup->resolution - mVec3s(1)) - mVec3f(mSmallest<float_t>((float_t)mMax(colourLookup->resolution))));
 
-  const float_t fracx = modf(pos.x, &pos.x);
-  const float_t fracy = modf(pos.y, &pos.y);
-  const float_t fracz = modf(pos.z, &pos.z);
+  const float_t fracx = modff(pos.x, &pos.x);
+  const float_t fracy = modff(pos.y, &pos.y);
+  const float_t fracz = modff(pos.z, &pos.z);
 
   const size_t x = (size_t)pos.x;
   const size_t y = (size_t)pos.y;
